Proposed modifications to tardis_kromer_plot.py
I am proposing new updates to tardis_kromer_plot.py, which now will only plot the elements in the model in the colormap. This leads to a much less cluttered colormap. Additionally, elements up to and including U have been added to enable trans Fe-group elements to be included. A new flag has also been introduced to enable the user to limit the number of elements to be included in the colormap. This is not necessary in most cases, but useful for models where many elements have been included, but most have negligible effects on the spectrum.