added in additional functionality
nelements flag is used to set the amount of elements to include in the colormap. If the no. of elements in the model exceeds this value, it ranks them based on their frequency of transitions, but this ranking of element frequency was performed across the entire model. The code has been modified to now do this within the wavelength range specified for the kromer plot