Update other species colouring
Previously the 'other species' category was created by stacking each of these species on top of each other. This led to some outlines appearing in the plot where they met. This update creates a single variable for hold the luminosity of the other species category so they are no longer stacked