Add capability to specify ions to be colored in Kromer Plot (#37)
* Update tardis_kromer_plot.py

In this update, I've added the ability to specify individual ions that you would like coloured in the Kromer plots, e.g. Si II. This is controlled through a new variable 'species_list'. Each species is given a unique 'species_id' variable that is then checked against the supplied list.

This update is a proof of concept and likely contains many bugs. In the current implementation, the species must be given in an ascending order (e.g. Si II, Si III, Ca II, etc. and not Ca II, Si III, Si II) to make sure the labels are correct. Each species must also be specifically request, e.g. Si II works, but not Si I - III, etc.

All other elements that are not requested are shown as silver, similar to the black and grey representing photosphere and e-scattering packets.

Is not currently compatible with elements, e.g. Si II, Ca, doesn't work but Si II, Ca II does.

* Added example Kromer plot showing species identification

* Fixed missing import

Added imports for species_string_to_tuple and species_tuple_to_string functions

* Added comments from James Gillanders

* Fixed formatting issues

* Added check for species notation

Check that all species are given in Roman numeral form, e.g. Si II and not Si 2 or Si 2+. If any species is in numerical form, raise an error

This removes ambiguity created due to numerical forms.

* Create .gitignore

* Added elements to species selection

Added ability to give an element in species list

* Added ion range

Added ability to specify ion range, e.g. Si I-III
Code will work out which ions fall within the request range and add them to the species list so that each one will get its own colour and label

* Update .gitignore

* Fixed colours

Fixed the colour mapping. If you only specify an element, it will colour all ions of that element the same

* Fixed bug when not specifying species list

Fixed an issue that meant the plots wouldn't work unless you specified a species_list. This is because it was expecting an ion id. If species_list is not specified, it will not default to the previous behaviour whereby it will colour elements the same

* Allow species to be given in random order

Implemented fix developed by James Gillanders that allows species to be given in any order, not just ascending

* Delete tardis_kromer_plot.cpython-36.pyc

* Fixed labels

Fixed issue with labels whereby they could be in the wrong order if a large number of ions are requested

* Fixed colouring for elements

Fixed issue that meant the colours weren't updated correctly if there was only one ion for a requested element

* Fixed colour bug

Fixed a bug for colours. If there were more than two ions present and an element was requested, not all ions would be coloured the same. Code should now be able to handle an arbitrary number of ions and give them all the same colour

* Added code comments

* Create ion_species_2.png

* Black formatting

* Fixed colour bug

Fixed issue that meant the colours weren't changed correctly for some different species.

* Fixed absorption colour bug

Same as previous commit, but now also applied to absorption

* Changed to is/is not None

* Changed comments to #

* Added description of nelements

* Update other species colouring

Previously the 'other species' category was created by stacking each of these species on top of each other. This led to some outlines appearing in the plot where they met. This update creates a single variable for hold the luminosity of the other species category so they are no longer stacked

* Updated within xlim selection

Implemented code developed by James Gillanders that is significantly simplified in how it works out the species within the xlims for absorption and emission.

Also required some restructuring of other bits of code. Specifically this is moving the nelements calculation outside of the line infos function

Co-authored-by: Mark Magee <mmagee37@qub.ac.uk>